backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
mysql-connector-python==9.0.0
python-dotenv==1.0.1
marshmallow==3.21.3
bcrypt==4.2.0
pytest==8.3.3

